splitting nodes into a module
