+++ psqlparse/nodes/__init__.py
@@ -1,0 +1,1 @@
+from .parsenodes import Statement
